+++ plugin/src/main/resources/io/github/mfederczuk/gradle/plugin/ktlint/git/hooks/other/pre-commit.template.sh
@@ -146,6 +146,9 @@ readonly ktlint_relative_opt_arg
 ktlint_limit_opt_arg=//KTLINT_LIMIT_OPT_ARG::quoted_string//
 readonly ktlint_limit_opt_arg
 
+ktlint_experimental_opt_arg=//KTLINT_EXPERIMENTAL_OPT_ARG::quoted_string//
+readonly ktlint_experimental_opt_arg
+
 ktlint_version=//KTLINT_VERSION::quoted_string//
 readonly ktlint_version
 
@@ -157,6 +160,7 @@ java -classpath "$ktlint_classpath" "$ktlint_main_class_name" \
      $ktlint_code_style_opt_arg \
      $ktlint_relative_opt_arg \
      $ktlint_limit_opt_arg \
+     $ktlint_experimental_opt_arg \
      --patterns-from-stdin='' < "$staged_kotlin_filename_list_file_pathname" ||
 	exc=$?